requirements.txt
ultralytics
opencv-python
streamlit
pillow
pyyaml
matplotlib
pandas
numpy
scikit-learn